address comments from #pq
