+++ src/io/flutter/editor/FlutterPubspecNotificationProvider.java
@@ -18,6 +18,7 @@ import com.intellij.ui.EditorNotificationPanel;
 import com.intellij.ui.EditorNotifications;
 import com.intellij.ui.HyperlinkLabel;
 import icons.FlutterIcons;
+import io.flutter.FlutterUtils;
 import io.flutter.pub.PubRoot;
 import io.flutter.sdk.FlutterSdk;
 import io.flutter.utils.FlutterModuleUtils;
@@ -56,11 +57,11 @@ public class FlutterPubspecNotificationProvider extends EditorNotifications.Prov
     }
 
     // Check that this pubspec file declares flutter
-    if (!PubRoot.declaresFlutter(file)) {
+    if (!FlutterUtils.declaresFlutter(file)) {
       return null;
     }
 
-    if(FlutterSdk.getFlutterSdk(project) == null) {
+    if (FlutterSdk.getFlutterSdk(project) == null) {
       return null;
     }
